MayoBot.java: Add skeletal greeting functionality
Chatbot only prints out a "hello" message along with a logo.

No functionality is implemented, and the Duke project is not differentiated from the template.

Implementing the basic skeletal functionality where the bot simple greets the user and exits is the first step to the breadth-first development approach for this project. Also, renaming it will allow for differentiation and adding personal touch.

Let's add basic functionality to the project where the bot introduces its name, greets the user, and exits. Let's also rename Duke to MayoBot to add personalization.