MayoBot.java: Add ability to store and display input text
The bot only repeats user input when entered, and only has the implementation of the exit command "bye".

Need to have some sort of memory of user input and begin implementing more commands for user to interact with the bot.

Let's add the ability for the bot to store whatever the user inputs, and implement the "list" command for the user to list out stored text.

Using a new TaskList for abstracting operations with regards to storing input text is preferable to maintain OOP principles and simplify implementation.