Task related classes: Add classes for Task and TaskList
Tasks are managed in the Task private static class within the main MayoBot class.

As the project expands, it would be preferable to manage Task and lists of Tasks separately in their own classes.

Let's add new class files Task.java and TaskList.java to manage these data objects and add corresponding operations to help with managing the code and commands of the main MayoBot chatbot.

Using separate classes is preferable to maintain OOP principles in the project.