MayoBot.java: Add ability to mark Tasks and done/not done
We can add input items to the bot and keep track of them, but no other operations on each item is possible. The main MayoBot class also does not implement the newly added task and TaskList classes.

We need to extend the functionality of the bot and add the ability for users to add and keep track of pending tasks.

Let's add the ability to mark tasks as done, and also add the ability to change the status back to "not done" by implementing the Task and TaskList classes.

Interfacing through the class abstraction barrier for Task and TaskList allows us to maintain OOP rpinciples in the project.