text-ui-test: Add tests using I/O redirection technique
The project has files and a script for automated text UI testing, but the input file is empty. The EXPECTED.txt file does not reflect expected outputs after recent iterations to the project.

In order to flesh out the project and ensure consistency and reliability in future iteration, testing needs to be implemented.

Let's implement the input.output redirection technique to semi-automate the testing of MayoBot, by adding input commands to input.txt, and adding the exected output in EXPECTED.txt.