MayoBot.java and Task-related classes: Add delete task functionality
Deletion of tasks is not supported by the chatbot.

Implementing deletion of tasks xtends the functionality of the bot.

Let's add support for deleting tasks from the list with the "delete" command, while also implementing its respective exception.

Alongside adding functionality for the "delete" command and its corresponding exception DeleteException, some refactoring of MayoBot.java and TaskList.java is needed to comply with OOP principles and encapsulation.