MayoBot: Add Parser and Command classes
The command parsing and execution logic is within the main MayoBot.java file.

In accordance with OOP principles, it would be more approrpiate to split and organize this logic into its own separate classes.

Let's create a Parser class that parses user command input, and creates a Command object. Let's transfer the runCommand() method in the main file to a statis method in the newly-created Command class.