Storage: Add Storage class to save tasks
The chatbot handles lists of tasks, but lists get erased every time the program is terminated.

Persistent storage of tasks saved in the list is needed to improve utility of the chatbot.

Let's begin to integrate the saving feature by adding a new Storage class to manage stored lists (in a .txt file) and support basic loading and saving operations.