Storage: Implement task marked/unmarked changes
The user is unable to save marked/unmarked task as done/not done in the local data file.

This limits the functionality of the task list saving feature, as user is not able to manipulate task entry after saving.

Let's allow the user to mark/unmark saved tasks by triggering updation of the saved tasks after the status of a task has been updated.