Storage: Allow task deletion to be saved
Any task deletion is not reflected in the user's local save data.

This limits the implementation of the task list saving feature, as it should be able to reflect all changes/manipulations made to the list.

Let's allow task deletions to be reflected in the save data, by triggering saveTasks() method whenever a task is deleted.