Ui and MayoBot: Add Ui class & simplify program logic
The main code for the bot exists in the main() method for the MayoBot program, with direct UI outputs and messy program logic. As of now, separate classes for Storage, Parser, Commands and TaskList exist.

In order to simplify code readability and adherence to OOP principles, extracting out closely related code as classes is necessary. One important functionality of the bot that needs to be extracted out into a class is UI.

Let's extract out all UI related methods into a separate Ui class, and improve the organization of the MayoBot main file by adding private attributes, consstructors, and a run() method that is called in the main() method.