Command class: Display messages through Ui class
When printing message to console, the Command.execute() method directly calls System.out.println() instead of working through the Ui class.

This violates separation of tasks and responsibilities rule in the code.

Let's add the showMessage() method to the Ui class and make it so that all messages triggered by Commands are printed to the ui through the Ui class.