Add JUnit tests for MayoBot core classes
The chatbot project relies on text input/output tests instead of relying on a more sophisticated test suite.

This limits the variety and complexity of tests that can be run for MayoBot.

Let's begin to add JUnit tests for the entire project by createst unit tests for the MayoBot core classes: MayoBot, Command, Parser, Storage, and Ui.