Storage, Ui, Command: Add JavaDoc comments
The chatbot code only has JavaDoc comments for the Parser class.

In line with the goal of improving JavaDoc documentation, we need to add these comments for the rest of the core function classes.

Let's continue this task by adding JavaDoc comments for Sotrage, Parser and Ui classes and their methods.