MayoBot and Task classes: Add JavaDoc comments
The project does not have JavaDoc comments for the main MayoBot class and all Task-related classes (inclduing TaskList).

As part of the goal of adding JavaDoc documentation, these comments have to be added to the rest of the classes involved in this project.

Let's do so by adding JavaDoc comments to the following classes: MayoBot, Task, TaskList, TodoTask, DeadlineTask, and EventTask.