Command and TaskList: Separate ui updates from findTask method
The findTask method in TaskList prints out corresponding output directly to the interface from the method itself.

To maintain separation of tasks, updates and output to UI should be through the Ui class.

Let's fix this within the "find" command implementation and the printTask/printTasks methods in TaskList by ensuring that outputs are displayed through the Ui class.

To make this change for Command#findTask, changes need to be made to Command#execute as well in the switch/case block for case "find".