Level-10: Add JavaFX GUI
The chatbot only has a CLI interface.

In order to enhance user experience and usability of the chatbot, having a GUI interface would be preferable.

Let's begin to add a GUI to MayoBot using JavaFX technology, by adding the needed classes, corresponding changes to exisying classes, and resources (such as CSS files and images).