MayoBot: Allow displaying of exceptions on GUI
The in-progress implementation of the JavaFX GUI does not support displaying of exceptions.

Displaying exceptions will allow user to use the chatbot more effectively and be guided on how to use comnmands correctly.

Let's allow display of exceptions on the GUI by returning the exception message in the MayoBot#getResponse() method.