Rename MayoBot package to mayobot and fix checkStyle
The main package for the project is named "mayoBot", which does not adhere to the style guide/convention for package names. Additionally, running checkStyles gives rise to multiple errors.

Adhering to style conventions is essential for the individual project.

Let's begin to resolve these style issues by renaming the MayoBot package to "mayobot", and doing necessary refactoring changes. Additionally, let us begin to resolve the checkStyle errors (expect the errors related to missing JavaDoc comments, those will be tackled in a future commit).