BetterGui: Polish GUI further
Although the GUI interface of the chatbot is good, there are some finishing touches left (such as background, color tweaks etc.)

Finalizing the GUI interface is important to prepare the project for release.

Let's make the needed GUI changes to the project by changung the background and some colors etc.